backend update, context update
Backends have a "Renderer.hs" in them now that contextually shows the use of the renderer and setup. The engine itself has a "Context.hs" module now that all Window management code is now in. This module is used to initialize a sdl2 window or a vulkan window instance and then plugin a backend rendering module to.