AAPLRenderer Class
an interface(in cbits/mtlpp.cc) that the backend actually uses to pass data to Apple M1 chipsets with. The AAPLRenderer class is all that needs to be interfaced with to render meshes, textures, and 3d data with. The Haskell interface then just makes calls to this class to render data with. Haskell interface reflections currently in progress.